Remove Code Duplication in Validation
Instead of hand-writing the code for Short and Int respectively, we now
convert the Short to Int and then call the Int implementation. There,
the value is checked to be inside the right Int range and then converted
to Short again to construct the class.

It is my understanding this does not change runtime behavior much
because the previous implementation also checked that the Short is
inside an Int range and probably also had to convert the Short to an
Int.