Make Patch Non-Singleton For Easier Testing
The fact that `Patch` is singleton was an issue during testing, as its
state always needed to be checked before being able to perform tests
that interact with it.

To solve this issue I changed Patch to a normal class that is
instantiated during the initialization of the CueGlowServer. The patch
must then be passed to the handler functions through some form of
dependency injection. I chose to structure it the following way:

The CueGlowServer keeps a state element (currently only wraps Patch, but
more will likely come in the future). The CueGlowServer also keeps an
InEventHandler which has access to the state through its constructor.

All Handler functions take a reference to the InEventHandler through
their method arguments. They call into the InEventHandler to produce
changes in the state.

This way we achieve abstraction between network related handlers and the
network-agnostic handlers (which now live in the InEventHandler).