Test and Implement updateFixture
The way to distinguish between a field in the message being absent and
a field being present with the value null now is a Result type. It only
works for parsing, not for serialization. It's also a lot of code, hard
to understand, adds even more converters, ...
I don't like it but don't know of a better solution.