Major Refactor of API Event Handling
Previously, GlowMessage was assumed to be an internal representation of
a message coming into the system, not matter which protocol it came
from. This is not a good idea, because GlowMessage was completely
littered with Klaxon annotations, converters and parsing code.

To ease adding other APIs in the future, I refactored according to the
system outlined in the file `cueglow-server/src/README.md`.

I refactored all GlowMessage, GlowEvent, GlowData and GlowRequest to
JsonMessage, JsonEvent, JsonData and JsonRequest. Whether that was the
best naming choice, I don't know.

But when we call it Glow-something, we indicate it is general over
different APIs and applies to the whole of CueGlow. What previously was
GlowMessage only applied to JSON and was never gonna apply to anything
else.

I left GlowError because it can be used generically throughout the whole
application to indicate domain errors.