Change Patch API to always allow multiple Elements
If the API for Patch allows multiple elements, the outgoing updates can
also be grouped which reduces overhead. Also, if all actions in JSON
API, Patch API and PubSub operate on a list of actions, everything is
more consistent.

The way error lists are returned from the Patch API isn't really great
yet, though. Also, the overhead of always spawning a list is a bit meh.
Could be reduced by refactoring and also allowing single-element
interaction with the API (make the multi-element methods use the
single-element methods).