Refactor JsonMessage to GlowMessage again
As discussed on TeamSpeak, we'll try to use the JSON-annotated
GlowMessage internally as well as for parsing to JSON. With more APIs in
the future, more isolation between internal and external APIs may be
needed and require a more clear separation between internal and external
representations. For now, we'll avoid the code duplication though.