Make Old JSON API Work (#32)
* Add Ability to Stop Server After Test

* Make Patch Non-Singleton For Easier Testing

The fact that `Patch` is singleton was an issue during testing, as its
state always needed to be checked before being able to perform tests
that interact with it.

To solve this issue I changed Patch to a normal class that is
instantiated during the initialization of the CueGlowServer. The patch
must then be passed to the handler functions through some form of
dependency injection. I chose to structure it the following way:

The CueGlowServer keeps a state element (currently only wraps Patch, but
more will likely come in the future). The CueGlowServer also keeps an
InEventHandler which has access to the state through its constructor.

All Handler functions take a reference to the InEventHandler through
their method arguments. They call into the InEventHandler to produce
changes in the state.

This way we achieve abstraction between network related handlers and the
network-agnostic handlers (which now live in the InEventHandler).

* Refactor Integration Tests

This way, the integration tests can be run on a server instance that is
only intitiated once.

Also includes code to output GDTF validation errors and tests for it.

* Intermediate state before refactoring working nextFreeFid

* Adding Fixtures kinda works

The WebSocket answer after adding does not work.

* Change GlowErrors to Classes with Reason

closes #23

* Test and Impelement Response to addFixtures

* Test Invalid addFixture Messages

* Refactor addFixtures Handler

* Remove Fixtures when Removing FixtureType

When a fixture type is deleted, now also all associated fixtures are
removed from the patch. Test modified to cover this feature.

* Ensure Invalid addFixture Does Not Add Fixtures

* Test and Implement updateFixture

The way to distinguish between a field in the message being absent and
a field being present with the value null now is a Result type. It only
works for parsing, not for serialization. It's also a lot of code, hard
to understand, adds even more converters, ...
I don't like it but don't know of a better solution.

* Test Different fixtureUpdate Functions

* Test and Implement deleteFixtures

* Test deleteFixtures with Invalid FixtureTypeId

* Documentation Improvements & Minor Changes

* Major Refactor of API Event Handling

Previously, GlowMessage was assumed to be an internal representation of
a message coming into the system, not matter which protocol it came
from. This is not a good idea, because GlowMessage was completely
littered with Klaxon annotations, converters and parsing code.

To ease adding other APIs in the future, I refactored according to the
system outlined in the file `cueglow-server/src/README.md`.

I refactored all GlowMessage, GlowEvent, GlowData and GlowRequest to
JsonMessage, JsonEvent, JsonData and JsonRequest. Whether that was the
best naming choice, I don't know.

But when we call it Glow-something, we indicate it is general over
different APIs and applies to the whole of CueGlow. What previously was
GlowMessage only applied to JSON and was never gonna apply to anything
else.

I left GlowError because it can be used generically throughout the whole
application to indicate domain errors.

* Move PatchFixtureData to JsonData and Rename

* Add API Documentation with new Pub/Sub API Design

* Make PatchFixture Immutable

This is part of making the Patch thread-safe.

* Change Patch API to always allow multiple Elements

If the API for Patch allows multiple elements, the outgoing updates can
also be grouped which reduces overhead. Also, if all actions in JSON
API, Patch API and PubSub operate on a list of actions, everything is
more consistent.

The way error lists are returned from the Patch API isn't really great
yet, though. Also, the overhead of always spawning a list is a bit meh.
Could be reduced by refactoring and also allowing single-element
interaction with the API (make the multi-element methods use the
single-element methods).

* Change WebSocket API from Javalin to Jetty

This allows us to spawn a stateful object for each WebSocket connection
that can each keep its own state for subscription, ping/pong, etc.

* Make WebSocketConnection implement AsyncClient directly

* Prefix WebSocket Classes with 'Glow' to differentiate from Jetty Names

* Change from WebSocketAdapter() to WebSocketListener for null safety

* Validate JsonMessage Data Type matches Event Type

* Refactor JsonMessage to GlowMessage again

As discussed on TeamSpeak, we'll try to use the JSON-annotated
GlowMessage internally as well as for parsing to JSON. With more APIs in
the future, more isolation between internal and external APIs may be
needed and require a more clear separation between internal and external
representations. For now, we'll avoid the code duplication though.

* Return 400 instead of 500 on GDTF Validation Fail

The Schema validation mostly checks the GDTF syntax without really
checking its semantic validity. Therefore, a status code 400 is
appropriate.

* Rename IncomingGlowMessageHandler

New Name: IncomingGlowRequestHandler

Reason: Does not handle GlowMessage but GlowRequest

* Refactor JSON-Specific Part into json Package

Now, an AsyncClient needs to accept a `GlowMessage` instead of a String
and must serialize the message itself.

This decouples the GlowRequest from any JSON-specific code.

* Instrument ApiIntegrationTest with Timing Code

Avg Reported time from start of ApiIntegrationTest to end of teardown:
3396 ms

Averaged over three runs of:
[3462.627858,
3343.020218,
3382.882273]

(times in ms)

* Restart Server for Each ApiIntegrationTest

Average Runtime: 3478 ms

over three iterations of:
[3446.472029,
3453.562753,
3534.502066]

(time in ms)

* Clean Up With Restarting Server

* Add Documentation for integration Test Package