JSON API Redesign (#35)
* Clarify GDTF API Response

* New AddFixtures API & New GlowMessage Polymorphism

Tests updated to new addFixtures API and passing. This required a
redesign of GlowMessage. Some cleanup in the GlowMessage Design may be
required a few commits down the road.

* New UpdateFixtures API

Includes many pieces of old code that were removed or cleaned up.

* New JSON Error API

Now, we directly bind GlowError to JSON. The parsing does not really
work, as an instance of `GlowError` is returned where name and
description fit, but not the specific class of the Error.

* New RemoveFixtures API

* New RemoveFixtureTypes API

* New fixtureTypeAdded API

* New PatchSubscribe API

Still fails test because of issue #24

* Update Klaxon and Make PatchSubscribe Test Pass

Klaxon: 5.4 -> 5.5

Use the new serializeNull feature to make the test for PatchSubscribe
pass.

* Don't Serialize Most Null Fields to JSON

Apply `@Json(serializeNull = false)` where applicable and test for it.

closes #24

* Remove GlowData

* Add PatchUnsubscribe

* Add New JSON addFixtureTypes

* Add PatchInitialState JSON Event