Allow Exiting out of NewFixtures with Esc
Normal HTML inputs don't blur on Esc, so it should be okay
to exit the page on Esc as announced by the Hotkey Label.

Possible Issue: A user assumes Esc will exit out of an input and
the user loses the data her already entered.
Since not much data can be entered in this form,
I think it is fine.