Fix Failing Tests in Chromium
This is a strange issue - clicking on the label and typing does work in
Firefox but not in Chromium. But the issue only shows up when
something is already typed into the field - it works for empty inputs.

Now we just directly type into the input associated with the label,
which works. The code for the utility function to do this is taken from
https://glebbahmutov.com/cypress-examples/6.5.0/recipes/form-input-by-label.html#reusable-function