Remove Frozen Values from State and Reverse Graph
I realized that keeping frozen values in disabled functions is
too complicated (the algorithm could easily deadlock, as each change
in value can cause a change in ModeMaster, which could again change
value, and so on).
I therefore removed it from the rig state. I also reversed the direction
of the depedency graph, to make the needed iteration easier.