Working Version with Channel Function Control over Art-Net (#63)
* First Draft for Describing GDTF Structure in CueGlow

* Fix Small Wording Issues in GDTF Doc

* Build and GDTF Doc Adjustments

* Remove Old Reference to sass-extract-loader

* Add List of MultiByteChannels to GlowDmxMode

* Parse GDTF ChannelFunctions Including DmxTo

* Move ChannelFunctions into Own Array

Previously, ChannelFunctions were part of the MultiByteChannel, making
it hard to understand a future value array with values for each
ChannelFunction.

Now, we link between MultiByteChannel and ChannelFunctions with array
indices.

* Add Raw DMX Channel Functions

This is a step towards handling Channel Function dependencies.

* Create Channel Function Dependency Graph

Includes some crude code for serialization of graph.
Unfortunately, the constructors for a DMX mode and its subtypes
become more and more convoluted. We'll need to refactor those later.

* LeftRightSplit Component for Main Page

* Add FixtureSelection Table to Main Window

* ChannelFunction Sliders and Show ChannelLayout

* Test ModeMaster Dependency Graph

Also includes the framework for handling the state of
GDTF Fixtures.

* Implement GDTF Default State

* Fix React Unique Key in Fixture Type Details

* Tracking in State Transitions

Changes in a channel due to changes in a ModeMaster
not implemented or tested yet.

* Failing Test for Nested Mode Master

The code for updating channels when a modeMaster
dependency of them changes is missing.

* Remove Frozen Values from State and Reverse Graph

I realized that keeping frozen values in disabled functions is
too complicated (the algorithm could easily deadlock, as each change
in value can cause a change in ModeMaster, which could again change
value, and so on).
I therefore removed it from the rig state. I also reversed the direction
of the depedency graph, to make the needed iteration easier.

* Implement and Test new State Transition

* Add Rig State to StateProvider

* Server-Side Messaging for RigState

* Fixture State Sync with Frontend

Generally works okay. You have to close your DevTools though,
otherwise performance is very bad.
There is still some UI bugginess when you delete fixtures in the
background or first access the fixture list.

* Implement ArtNetSending

Still TODO: Double Packets during Suppression

* Fix UI Crash when Accessing Freshly Added Fixture

* Make Locked Code Portion in ArtNetSending Smaller

It allocates a lot now, though. Not sure if it performs better, but I
guess it's less nested in terms of formatting.

* Remove Hardcoded ArtNet Target Address

Replaced by global broadcast 255.255.255.255, which should just send to
all in the local network. Not best practice, but for this application
probably fine.

* Fix Padding Issues with ChannelFunction Sliders

* Show ChannelFunction Geometry and Attribute

Fix some Backend Tests while we're at it

* Sort ChannelFunctions by FeatureGroup

* Show DMX Offsets in ChannelFunctions

* Display Associated DMXChannel Name

* Show outOfRange and disabled Tooltips

* Prevent Crash of ArtNet Thread

You could for example cause that by patching a fixture at such a high
address that it would create values beyond address 512.

We still should do some other kind of validation, but now at least the
ArtNet output starts back up again if something goes wrong.

* Handle REST Error when Uploading GDTF

Now displays a toast with the error.

* roughly Optimize Slider Performance

Memoizing the Sliders with unchanged values does the trick.
Also moving the Context Provider down to where it's necessary.
The code still needs some refactors for readability.

* Refactored Memoization

I find this new one more readable, but I'm not sure performance
is as good as it was before.

With the current version and a production build, I drop a
frame every few frames.

* Fix Tables Sometimes not Updating

* Simple Patch Validation and Default Address

Also, ArtNetSending now truncates the end of oversized DMX output.

* Fix Default Address in New Fixture UI

closes #22

* Add GDTF Test for Cyclic Dependencies

* Simple Cypress Test for Channel Function Slider

Also, New Fixture now counts up the address of multiple fixtures.
This was needed to pass the Cypress tests with the new Patch
Validation Message and is a good idea anyway.

* Multi-Column Layout for Channel List

* Show GDTF Metadata

* Send a Ping Every Minute

This prevents the WebSocket Timeout in Jetty when there
is no interaction.

* First Try at ModeMaster Dependency Graph

* Clean Up ChannelFunctionGraph

* Move ModeMaster Graph to its own Popup

This gives it more space, which is needed for the graphs to be
readable.

* Refine ModeMaster Graph Layout

* Try to Fix Ping Interval Initilization

I saw the WebSocket being automatically disconnected,
so maybe it had something to do with the event loop
being broken between the initializations?

* Fix Keyboard Behavior with Overlays

closes #57

* Change RigState from Fixture Index to UUID

Unfortunately, I also did a reformat on the whole Kotlin project.
That's why there are even more changed files than necessary.

* Fix Bug where Tables Don't Update

* Hide ModeMaster Graph Button When Graph Empty

* Enfroce New Fixture Adresses Are Inside Universe

* Exit When Hitting Escape in New Fixture Inputs

* Private Constructor for  ArtNetAddress/DmxAddress

There was a comment saying Klaxon required the constructor to be
public, but I made it private and everything works.

closes #41

* Update Server Dependencies

* Disable npmInstall in runNoNpm

* Run npm update

* Update Cypress

* Inconsequential Frontend Updates

* Update @reach/router to react-router v6

This will allow us to update to React 17.

* Update react-scripts and Tabulator

- The new version of react-scripts uses /ws for its hot reloading, so I
changed our API to `/webSocket`. The server doesn't care anyway.
- I added the sass package, which we need now that node-sass is removed
- I removed react-tabulator because it would not allow us to update
React. I replaced it with my own wrapper of the new
Tabulator v5. Since the types aren't quite correct yet,
I patched the types with patch-package.

* Update to React 17

Also fix Linting Problems in GlowTabulator

* Renaming to GlowDTF and Documentation Improvements (#65)

* Draft for GlowDTF README and CONTRIBUTING

* Remove CueGlow Name from Folders

Besides renaming almost every file, I also had to change some values in
`server/npm.gradle` to make it work.

Left ToDo: Adjust Documentation

* Remove CueGlow Name from Documentation

* Remove CueGlow Name from the Code

* Security Notice in README

* Word Wrap in DMX Channels

This may break channel names at non-whitespace characters
but at least they don't overlap each other now.

* Delete Old Blueprint CSS Variables

* Auto-Increment FID for Multiple Fixtures

* Make ChannelFunction Slider Padding Symmetric

* Set Up Fat Jar Packaging

* Disable Logging to File

I think it is more convenient for users to not have to clean
up our logs.

* Clarify README and Add GDTF Resources

* Add User Guide to README

* Add Developer Docs Hint to README